# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reddint API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/comments": {
      "post": {
        "summary": "Create a new Comment",
        "tags": [
          "Comment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/CreateComment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in!"
          },
          "404": {
            "description": "Wrong credentials."
          }
        }
      }
    },
    "/comments/{post_id}": {
      "get": {
        "summary": "Get all comments of a post",
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "schema": {
              "type": "string",
              "required": true
            }
          }
        ]
      }
    },
    "/comments/upvote/{comment_id}": {
      "put": {
        "summary": "Upvote a comment",
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "schema": null,
            "type": "string",
            "required": true
          }
        ]
      }
    },
    "/comments/downvote/{comment_id}": {
      "put": {
        "summary": "Downvote a comment",
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "schema": null,
            "type": "string",
            "required": true
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateComment": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Comment body"
          },
          "post_id": {
            "type": "string",
            "description": "Post id where the comment belongs"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "example": {
          "body": "🎉 Your post sucks 🎉",
          "post_id": "5e9f8f8f8f8f8f8f8f8f8f8"
        }
      },
...
    }
  },
...
}